Mms: Fix recipient text can't scroll to first
The recipient address input can't scroll to the first.
When the selection was changed, it will set the selection back
to the end of text.
In on touch event, it will set the selection to the end of text.

When user is scroll, set a flag, and when user leave the screen,
reset this flag. When this flag is set, didn't handle the touch
event and the selection change event.

CRs-fixed: 525934

Change-Id: I1cd9bb0f7cc7ba2c2061e6728bf445b5dd6985f2