MMS: Fix new message can't be saved as draft after press Back key.
If RecipientsEditor can't pending the chips correctly, the recipient
information will be lost, so the thread will be lost too.
There are two kind of situation when RecipientsEditor pending the chips.
One is chips < CHIP_LIMIT, another is chips > CHIP_LIMIT. But the
chips = CHIP_LIMIT situation is forgoten. So we need to deal with it.

Make sure when RecipientsEditor pending the chips, all situation will be
considered.

CRs-Fixed: 534081

Change-Id: I0bfa2d02440018865e1c601b75b9c6cb1bed99bb