MMS: Fix can't cut or copy text under landscape mode
In portrait mode, it will not trigger the select action when long
press the text in RecipientEdit. But under landscape mode, the IME
will process the long click event and start select action mode.
When selection changes, and it is inside the chips area. RecipientEdit
move the cursor back after the chips again. So selection start is the
same as selection end. It will do nothing when press cut or copy
text button.

Do not move the cursor back after the chips in full mode and remove the
span when user deleted some text OR some text was replaced.

CRs-Fixed: 539204

Change-Id: I4eb71e448a745c6d2d1272fd6473cb11e4ab9292