MMS: Fix auto match contacts when manual input recipient in new SMS failed
After RecipientEdit lost focus, chips after the second will be removed
temporary. When RecipientEdit get focus again, all removed chips will be
re-add. When re-adding chips, it use the chip's OriginalText to calculate
span end. But the OriginalText has been removed the last comma(',') during
createReplacementChip. So when calculating the span end, the length of ','
should be considerd, which means + 1 is necessary.

Set the right span end when re-add the spans that were removed.

CRs-Fixed: 543729

Change-Id: Ic865d483d8645df731942143fc673ff4df610edf