Mms: Contact in recipient box turns from name to number
When tap a contact Chip and press "backspace" key in virtual
keyboard to delete it, onTextChanged() will be called twice:
(1). "before - count == 1" will be true and the current Chip
(with Span and Text) is deleted. (2). "before > count" will be
true and only Span of next Chip will be deleted. Thus, the
next contact Chip will turn to numbers.

For design, "before > count" branch only can work in the full
screen mode of input method. When it is normal mode, we should
ignore this branch.

CRs-Fixed: 627988

Change-Id: I82c6f6928ab6328993661736955d62753ecc2b24